requirements.txt
vizro
vizro-ai
dash
flask
pandas
python-dotenv
docker